MODPATBLK-122 Rename LogHelper
