+++ src/main/java/org/folio/repository/BaseRepository.java
@@ -1,7 +1,7 @@
 package org.folio.repository;
 
-import static org.folio.util.LogHelper.logAsJson;
-import static org.folio.util.LogHelper.logList;
+import static org.folio.util.LogUtil.logAsJson;
+import static org.folio.util.LogUtil.logList;
 
 import java.util.List;
 import java.util.Optional;
