+++ src/main/java/org/folio/rest/impl/AutomatedPatronBlocksAPI.java
@@ -2,9 +2,9 @@ package org.folio.rest.impl;
 
 import static io.vertx.core.Future.succeededFuture;
 import static java.lang.String.format;
-import static org.folio.util.LogHelper.logAsJson;
-import static org.folio.util.LogHelper.logOkapiHeaders;
-import static org.folio.util.LogHelper.loggingResponseHandler;
+import static org.folio.util.LogUtil.logAsJson;
+import static org.folio.util.LogUtil.logOkapiHeaders;
+import static org.folio.util.LogUtil.loggingResponseHandler;
 import static org.folio.util.UuidUtil.isUuid;
 
 import java.util.Map;
