+++ src/main/java/org/folio/util/LogUtil.java
@@ -23,14 +23,14 @@ import io.vertx.core.buffer.Buffer;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.web.client.HttpResponse;
 
-public class LogHelper {
-  private static final Logger log = LogManager.getLogger(LogHelper.class);
+public class LogUtil {
+  private static final Logger log = LogManager.getLogger(LogUtil.class);
   public static final String R_N_LINE_SEPARATOR = "[\\r\\n]";
   public static final String R_LINE_SEPARATOR = "\\r";
   private static final int MAX_OBJECT_JSON_LENGTH = 10 * 1024;
   private static final int DEFAULT_NUM_OF_LIST_ELEMENTS_TO_LOG = 10;
 
-  private LogHelper() {
+  private LogUtil() {
     throw new IllegalStateException("Utility class");
   }
 
@@ -76,7 +76,7 @@ public class LogHelper {
             : format("first %d element%s", numberOfElementsToLog,
             numberOfElementsToLog == 1 ? "" : "s"),
           list.subList(0, numberOfElementsToLog).stream()
-            .map(LogHelper::logAsJson)
+            .map(LogUtil::logAsJson)
             .collect(Collectors.joining(", ")));
       }
     } catch (Exception ex) {
